+++ atlas-content-modeler.php
@@ -64,3 +64,12 @@ function atlas_content_modeler_loader(): void {
 	acm_content_connect_autoloader();
 	\WPE\AtlasContentModeler\ContentConnect\Plugin::instance();
 }
+
+/**
+ * Determines if opt-in usage tracking is enabled.
+ *
+ * @return bool True if enabled, false if not.
+ */
+function acm_usage_tracking_enabled(): bool {
+	return (bool) get_option( 'atlas_content_modeler_usage_tracking', false );
+}