+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Atlas Content Modeler Changelog
 
 ## Unreleased
+### Added
+- Opt-in anonymous usage tracking to help us make Atlas Content Modeler better (disabled by default).
 ### Fixed
 - Ensure Rich Text fields load for publishers even if WordPress Core editor scripts are slow to execute.